[otbn,dv] Add coverage tracking for loop stack
This uses the `COVER macro in the bound-in otbn_loop_if rather than
pushing everything all the way to the otbn_env_cov class. We want to
use sequences for the temporal logic ("see this, then that") and it
seems simpler to do it this way, rather than plumb absolutely
everything through sequence items.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>