[otbn,dv] Fix early exits in otbn_base_vseq::run_otbn() on reset
If the block gets reset, we want run_otbn() to exit immediately. In
practice, this already happens (because csr_wr and csr_spinwait both
exit immediately). But we make the early exit explicit to avoid
checking the end address: that check will fail if there is a reset in
the middle of the first execution of a binary.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>