[sw] Pass on native C/C++ compiler to Mundane build
In systems with more than one compiler users can choose the native (x86)
C/C++ compiler to use e.g. with the following command:

```
CC_FOR_BUILD=gcc-10 CXX_FOR_BUILD=g++-10 ./meson_init.sh -f
```

This compiler is then tested and used for native builds triggerd from
within Meson. Additionally, we also spawn out to other build systems,
and they need to know which compiler to use as well.

In particular, we build the Mundane Rust crate, which itself calls CMake
to build BoringSSL ("native", i.e. for the build machine).

With this change, we pass on the native compiler in a way that Mundane
gets ultimately built with the same native compiler as the rest of our
x86 ("native") software.

(This came up when trying to switch to an older GCC version due to
Mundane being unable to build with GCC 11, see
https://github.com/lowRISC/opentitan/issues/6978.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>