[tlul] Drop the "quick fail" path from tlul_adapter_sram.sv
Before this change, the adapter responded immediately on the A channel
if the request would cause an internal error (assuming there was space
in the request FIFO). The logic is that there's no need to wait for a
grant from the SRAM side since we're not sending anything to the SRAM
anyway.

This patch removes that logic so now the TL interface is stalled
unconditionally if !gnt_i. This might shorten some timing paths a
bit (since we no longer have the combinatorial path from request to
response, unless there's one on the SRAM side). However, my actual
motivation is to be able to assert things like "if gnt_i is false, the
TL side will stall" without having to add special handling for the
error case.

There will be a small observable slow-down for responses that should
generate an error, but this doesn't seem like a big deal (it's an
error! We probably don't care how quickly it's reported!). Indeed,
making that sort of thing more uniform seems like a more secure
approach anyway since it might eliminate some timing channel.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>