[lc_ctrl/otp_ctrl] Combine escalation actions 1 and 2
This simplifies the escalation actions by combining "wipe secrets"
(which is fanned out into the system via lc_escalate_en) and "lc state
scrap". This makes implementation and DV easier, since it is difficult
to logically keep these two actions apart without introducing a whole
set of unnecessary corner cases. For example, lc_escalate_en basically
moves all OTP partition controllers into an invalid state. This in turn
scraps the life cycle state automatically - however, it that one is
supposed to be a separable action, we need to introduce more logic to
actually handle that case. In reality, we are already recommending to
trigger both actions at the same time in the alert handler, so we do not
really need this separation.

In terms of implementation, the escalation connections are kept in place
and are just renamed. That means that we are going to have two
escalation channels going over to life cycle, and both of them trigger
the same response (i.e., they become redundant channels). This is not
unreasonable from a security perspective, and we need to connect the
unused channel somehow anyways to keep the ping mechanism alive.

Signed-off-by: Michael Schaffner <msf@opentitan.org>