[reggen] Remove swaccess and hwaccess from Register objects
In the hjson format, you can specify swaccess and hwaccess for a
register. This is just a handy shortcut to avoid having to specify it
for all the constituent fields. It's not supposed to mean anything
further. Remove swaccess and hwaccess from the Register class to make
sure no-one is (mis)using them.

Some knock-on effects are:

  - The REGWEN checks in RegBlock.validate() are now a bit stronger,
    checking that there's exactly one field in a REGWEN register,
    covering bit 0. (And then using that field for the access checks
    that existed before)

  - The checks for sensible access types for hwext registers now get
    applied to the fields, rather than the register itself.

  - The list of registers that aren't HW-writable that we collect for
    FPV (hro_regs_list in fpv_csr.sv.tpl) now makes sure not to
    include any registers where some (but not all) fields are HW
    writable.

I've checked, and this patch has no effect on the generated RAL
model (with topgen.py -t).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>