[otbn,dv] Teach the RIG about the 'wrd' operand of bn.mulqacc.so
This is both a source and a destination (we update just half of the
word). So we need to teach the tooling about updated wide registers (I
used the mnemonic "wrb") and update bignum-insns.yml to say that
bn.mulqacc.so's "wrd" operand is actually updated.

This fixes a failing test (otbn_reset with seed 531773529) caused by
the fact that one of the WDRs comes up with a nonzero initial value.
The RIG is quite careful to avoid reading registers with unknown
values, but didn't realise that a bn.mulqacc.so wasn't safe (because
the destination register didn't have a determined value).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>