[dv,tl] Read synchronously from TL agents in cip_base_env
This means that any scoreboard that derives from the cip_base_scoreboard
knows that its process_tl_access() task will be called in a
"synchronous manner". Specifically, if there are transactions on both
the A and D channels at the same time, it will see the D side and then
the A side.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>