[dv,cip] Mirror FATAL_ALERT_CAUSE-style registers in scoreboard
When running the TL error sequence run by run_tl_intg_err_vseq_sub, we
inject an integrity error into one or more TL accesses, which would
confuse the scoreboard for some blocks. To avoid chaos, we disable the
scoreboard when running this sequence with a runtime plusarg.

Some blocks' environments do prediction of hardware-updated CSRs, and
this is done in the scoreboard. Unfortunately, disabling the
scoreboard means that this won't happen which, in turn, causes
csr_utils::csr_rd_check to spot an "error" when reading a
hardware-updated CSR that changes on an alert (an "ALERT_CAUSE"
register).

This patch allows a block to configure the (single-bit) field that
will be set when there is a TL error. Assuming there's such a field
configured, cip_base_scoreboard.sv notices any error response and
updates the mirrored value of that field accordingly.

This allows us to get rid of the special-case code at the bottom of
cip_base_vseq__tl_errors.svh: rather than avoiding looking at a
particular behaviour, we can just predict it properly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>