[lib] Add launder32()
This function adds a "value laundering" pessimization that can be used
in targeted scenarios where the compiler's analysis (typically GVN or
something like it) would fold away critical check-your-work operations,
such as double-branches or loop completion checks.

The implementation is the result of careful analysis of existing
solutions and consultation with lowRISC's toolchain engineers. Usage of
this "intrinsic" is very subtle and requires careful review and
documentation.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>