[otp_ctrl] Remove invalid command error
This changes the behavior of the MacroError (code 0x1).

In particular, it is not asserted anymore when an invalid command
is issues to the wrapper.

Such a condition would be a HW bug, and should not exist in the final
design.

Hence, the behavior of that error is changed such that it is only
returned upon macro malfunction.

The open source OTP wrapper does not model this condition, hence the
corresponding test is removed.
I.e., this condition will have to be checked as part of the closed
source wrapper DV.

In order to catch illegal HW behavior, this change adds assertions that
check whether the otp controller issues legal commands to the OTP
wrapper.

Signed-off-by: Michael Schaffner <msf@opentitan.org>